SH-1 updates
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@1142 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/sh/include/types.h
@@ -71,7 +71,7 @@ typedef unsigned long long uint64;
  * irqsave()
  */
 
-typedef unsigned int irqstate_t;
+typedef unsigned long irqstate_t;
 
 #endif /* __ASSEMBLY__ */
